+++ ChangeLog.md
@@ -1,5 +1,9 @@
 # Changelog for uplift
 
+## Unreleased changes
+### Added
+- Add methods for uplift modeling from separate labels.
+
 ## 0.2.0
 ### Added
 - Add methods for uplift modeling from joint labels